authorization implemented

+++ src/api/notes/handler.js
@@ -12,7 +12,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'Untitled', tags, body } = request.payload;
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId });
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil ditambahkan',
@@ -43,8 +45,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -54,8 +57,10 @@ class NotesHandler {
   }
 
   async getNoteByIdHandler(request, h) {
-    const { id } = request.params;
     try {
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -85,8 +90,10 @@ class NotesHandler {
   }
 
   async putNoteByIdHandler(request, h) {
-    const { id } = request.params;
     try {
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       this._validator.validateNotePayload(request.payload);
       await this._service.editNoteById(id, request.payload);
       return {
@@ -115,8 +122,10 @@ class NotesHandler {
   }
 
   async deleteNoteByIdHandler(request, h) {
-    const { id } = request.params;
     try {
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
       return {
         status: 'success',
@@ -125,7 +134,7 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'success',
+          status: 'fail',
           message: error.message,
         });
         response.code(error.statusCode);
